现在经理可以对已完结课程进行结算

+++ src/main/java/base/model/InstitutionExample.java
@@ -314,52 +314,52 @@ public class InstitutionExample {
             return (Criteria) this;
         }
 
-        public Criteria andBalanceEqualTo(Integer value) {
+        public Criteria andBalanceEqualTo(Double value) {
             addCriterion("balance =", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceNotEqualTo(Integer value) {
+        public Criteria andBalanceNotEqualTo(Double value) {
             addCriterion("balance <>", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceGreaterThan(Integer value) {
+        public Criteria andBalanceGreaterThan(Double value) {
             addCriterion("balance >", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceGreaterThanOrEqualTo(Integer value) {
+        public Criteria andBalanceGreaterThanOrEqualTo(Double value) {
             addCriterion("balance >=", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceLessThan(Integer value) {
+        public Criteria andBalanceLessThan(Double value) {
             addCriterion("balance <", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceLessThanOrEqualTo(Integer value) {
+        public Criteria andBalanceLessThanOrEqualTo(Double value) {
             addCriterion("balance <=", value, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceIn(List<Integer> values) {
+        public Criteria andBalanceIn(List<Double> values) {
             addCriterion("balance in", values, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceNotIn(List<Integer> values) {
+        public Criteria andBalanceNotIn(List<Double> values) {
             addCriterion("balance not in", values, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceBetween(Integer value1, Integer value2) {
+        public Criteria andBalanceBetween(Double value1, Double value2) {
             addCriterion("balance between", value1, value2, "balance");
             return (Criteria) this;
         }
 
-        public Criteria andBalanceNotBetween(Integer value1, Integer value2) {
+        public Criteria andBalanceNotBetween(Double value1, Double value2) {
             addCriterion("balance not between", value1, value2, "balance");
             return (Criteria) this;
         }
